requirements.txt
# Kali SSE MCP 命令执行器依赖包

# MCP 协议支持
fastmcp>=0.9.0
mcp>=1.0.0

# Web 框架和 SSE 支持
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.5
websockets>=12.0

# 异步支持
asyncio-mqtt>=0.16.0
aiofiles>=23.2.1
aiohttp>=3.9.0

# 安全和加密
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 数据处理和验证
pydantic>=2.5.0
pydantic-settings>=2.1.0
marshmallow>=3.20.0
jsonschema>=4.20.0

# 日志和监控
structlog>=23.2.0
prometheus-client>=0.19.0
psutil>=5.9.6

# 数据库和缓存
sqlalchemy>=2.0.23
alembic>=1.13.0
redis>=5.0.1
sqlite3

# 测试框架
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.2

# 开发工具
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# 系统工具
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# 网络和HTTP
requests>=2.31.0
urllib3>=2.1.0

# 配置管理
python-dotenv>=1.0.0
toml>=0.10.2
pyyaml>=6.0.1

# 进程管理
supervisor>=4.2.5

# 时间处理
python-dateutil>=2.8.2
pytz>=2023.3

# 正则表达式增强
regex>=2023.10.3

# 文件处理
pathlib2>=2.3.7
watchdog>=3.0.0

# 性能分析
memory-profiler>=0.61.0
line-profiler>=4.1.1

# 文档生成
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# API 文档
fastapi-users>=12.1.2
python-jose[cryptography]>=3.3.0

# 环境变量管理
python-decouple>=3.8

# 数据序列化
orjson>=3.9.10
msgpack>=1.0.7

# 网络扫描工具接口
python-nmap>=0.7.1
shodan>=1.30.1

# 命令行解析增强
argparse>=1.4.0
shlex

# 系统信息
distro>=1.8.0
platform

# 并发处理
concurrent.futures
threading
multiprocessing

# 错误处理和重试
tenacity>=8.2.3
backoff>=2.2.1

# 数据分析（用于智能化功能）
pandas>=2.1.3
numpy>=1.25.2
scikit-learn>=1.3.2

# 机器学习（用于错误学习）
joblib>=1.3.2

# 图形和可视化（用于报告生成）
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# 模板引擎
jinja2>=3.1.2

# XML/HTML 处理
lxml>=4.9.3
beautifulsoup4>=4.12.2

# 图像处理（用于报告）
pillow>=10.1.0

# 地理位置（用于IP定位）
geoip2>=4.7.0
maxminddb>=2.2.0

# 缓存装饰器
cachetools>=5.3.2

# 配置验证
cerberus>=1.3.5

# 进度条
tqdm>=4.66.1

# 颜色输出
colorama>=0.4.6
termcolor>=2.3.0

# 系统监控
py-cpuinfo>=9.0.0

# 网络工具
netaddr>=0.9.0

# 时间工具
arrow>=1.3.0

# 文件锁
filelock>=3.13.1

# 装饰器工具
decorator>=5.1.1

# 类型提示
typing_extensions>=4.8.0

# 版本比较
packaging>=23.2

# 调试
debugpy>=1.8.0

# 代码覆盖率
coverage>=7.3.2

# 静态分析
pylint>=3.0.3
bandit>=1.7.5

# 安全扫描
safety>=2.3.5

# 依赖分析
pipdeptree>=2.13.1

# 虚拟环境
virtualenv>=20.25.0

# 构建工具
build>=1.0.3
twine>=4.0.2

# 版本控制
gitpython>=3.1.40

# SSL/TLS
certifi>=2023.11.17
